adding the viewHolder in the activity due to a problem
