+++ app/src/main/java/fr/bigsis/android/activity/TripListActivity.java
@@ -9,6 +9,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 import android.widget.ImageButton;
+import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -24,6 +25,8 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
+import com.bumptech.glide.Glide;
+import com.bumptech.glide.request.RequestOptions;
 import com.firebase.ui.firestore.paging.FirestorePagingAdapter;
 import com.firebase.ui.firestore.paging.FirestorePagingOptions;
 import com.firebase.ui.firestore.paging.LoadingState;
@@ -31,6 +34,9 @@ import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
+
+import java.text.SimpleDateFormat;
+import java.util.Locale;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -41,7 +47,6 @@ import fr.bigsis.android.fragment.SearchMenuFragment;
 import fr.bigsis.android.fragment.ToolBarFragment;
 import fr.bigsis.android.helpers.KeyboardHelper;
 import fr.bigsis.android.view.CurvedBottomNavigationView;
-import fr.bigsis.android.viewHolder.TripListViewHolder;
 import fr.bigsis.android.viewModel.SearchMenuViewModel;
 
 
@@ -121,8 +126,10 @@ public class TripListActivity extends AppCompatActivity implements SearchMenuFra
     public void onFragmentInteraction() {
         onBackPressed();
     }
+
     private void setToolBar() {
         transitionContainer = (ConstraintLayout) findViewById(R.id.toolbarLayout);
+        transitionContainer.setBackground(getDrawable(R.drawable.gradient));
         imbtSearch = (ImageButton) transitionContainer.findViewById(R.id.imBt_search_frag);
         imBtAdd = (ImageButton) transitionContainer.findViewById(R.id.imBt_add_frag);
         imBtCancel = (ImageButton) transitionContainer.findViewById(R.id.imBt_cancel_frag);
@@ -190,6 +197,42 @@ public class TripListActivity extends AppCompatActivity implements SearchMenuFra
                 .commit();
     }
 
+    public class TripListViewHolder extends RecyclerView.ViewHolder {
+
+        @BindView(R.id.tvTripsFrom)
+        TextView mTextFrom;
+
+        @BindView(R.id.tvTripsTo)
+        TextView mTextTo;
+
+        @BindView(R.id.ivTripImage)
+        ImageView mImvTripImage;
+
+        @BindView(R.id.tvDateTrip)
+        TextView mTextDate;
+
+        public TripListViewHolder(@NonNull View itemView) {
+            super(itemView);
+            ButterKnife.bind(this, itemView);
+        }
+
+        public void bind(@NonNull TripEntity item) {
+            mTextFrom.setText(item.getFrom());
+            mTextTo.setText(item.getTo());
+            SimpleDateFormat format = new SimpleDateFormat("E dd MMM, HH:mm", Locale.FRENCH);
+            mTextDate.setText(format.format(item.getDate().getTime()));
+
+            RequestOptions myOptions = new RequestOptions()
+                    .fitCenter()
+                    .override(250, 250);
+
+            Glide.with(mImvTripImage.getContext())
+                    .asBitmap()
+                    .apply(myOptions)
+                    .load(item.getImage())
+                    .into(mImvTripImage);
+        }
+    }
     private void setUpAdapter() {
         Query baseQuery;
